Feat : 유저 필수정보 입력 컨트롤러 작성 및 컨트롤러 정리
