--- src/main/java/com/projectmatching/app/domain/user/dto/UserTechDto.java
@@ -1,13 +1,0 @@
-//package com.projectmatching.app.domain.user.dto;
-//
-//import com.fasterxml.jackson.annotation.JsonInclude;
-//import lombok.*;
-//
-//@Getter
-//@Setter
-//@AllArgsConstructor(access = AccessLevel.PRIVATE)
-//@NoArgsConstructor
-//@Builder
-//@JsonInclude(JsonInclude.Include.NON_NULL) //null 이면 생성되지 않음
-//public class UserTechDto {
-//}
